Reorganize code and update tests
